requirements.txt
# Funding Round Data Collection Engine V2 - Dependencies

# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Search
ddgs>=9.0.0  # DuckDuckGo search

# Data processing and export
pandas==2.1.3
openpyxl==3.1.2

# Retry logic
tenacity==8.2.3

# Logging
coloredlogs==15.0.1
